[UPDATE] Admin permission for planets create

+++ Ogame/Ogame/Controllers/PlanetsController.cs
@@ -65,8 +65,14 @@ namespace Ogame.Controllers
         }
 
         // GET: Planets/Create
-        public IActionResult Create()
+        public async Task<IActionResult> Create()
         {
+            User user = await GetCurrentUserAsync();
+            if (!user.IsAdmin)
+            {
+                return NotFound();
+            }
+
             ViewData["UserID"] = new SelectList(_context.Users, "Id", "Id");
             return View();
         }
@@ -78,6 +84,12 @@ namespace Ogame.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("PlanetID,UserID,Name,Dist_to_star,X,Y,Metal,Cristal,Deuterium,Energy")] Planet planet)
         {
+            User user = await GetCurrentUserAsync();
+            if (!user.IsAdmin)
+            {
+                return NotFound();
+            }
+
             if (ModelState.IsValid)
             {
                 _context.Add(planet);
@@ -92,6 +104,12 @@ namespace Ogame.Controllers
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
+            {
+                return NotFound();
+            }
+
+            User user = await GetCurrentUserAsync();
+            if (!user.IsAdmin)
             {
                 return NotFound();
             }
@@ -113,6 +131,12 @@ namespace Ogame.Controllers
         public async Task<IActionResult> Edit(int id, [Bind("PlanetID,UserID,Name,Dist_to_star,X,Y,Metal,Cristal,Deuterium,Energy")] Planet planet)
         {
             if (id != planet.PlanetID)
+            {
+                return NotFound();
+            }
+
+            User user = await GetCurrentUserAsync();
+            if (!user.IsAdmin)
             {
                 return NotFound();
             }
@@ -149,14 +173,19 @@ namespace Ogame.Controllers
                 return NotFound();
             }
 
+            User user = await GetCurrentUserAsync();
+            if (!user.IsAdmin)
+            {
+                return NotFound();
+            }
+
             var planet = await _context.Planets
                 .Include(p => p.User)
                 .FirstOrDefaultAsync(m => m.PlanetID == id);
 
-            User user = await GetCurrentUserAsync();
-            if (planet == null || planet.UserID != user.Id)
+            if (planet == null)
             {
-                return RedirectToAction(nameof(Index));
+                return NotFound();
             }
 
             return View(planet);
@@ -167,6 +196,12 @@ namespace Ogame.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            User user = await GetCurrentUserAsync();
+            if (!user.IsAdmin)
+            {
+                return NotFound();
+            }
+
             var planet = await _context.Planets.FindAsync(id);
             _context.Planets.Remove(planet);
             await _context.SaveChangesAsync();
